Edit crypto (+3)
- Rename mod stream to io
- Add struct Chain to mod io
- Crypto now takes Read and Write trait implementors
