+++ src/lib/lib.rs
@@ -15,9 +15,9 @@
 pub mod constants;
 pub mod crypto;
 pub mod file_header;
+pub mod io;
 pub mod metadata;
 pub mod pack;
-pub mod stream;
 pub mod tmp;
 pub mod walk_dir;
